+++ lib/omeka_client.rb
@@ -3,14 +3,17 @@ require "json"
 require "omeka_client/version"
 require "omeka_client/client"
 require "omeka_client/omeka-item"
+require "omeka_client/omeka-site"
+require "omeka_client/omeka-collection"
+require "omeka_client/dublin-core"
+require "omeka_client/item-type-metadata"
 
-
-# 
+#
 # The module wrapping the classes that do the work
-# 
+#
 # @author Lincoln Mullen
 # @since  0.0.1
-# 
+#
 module OmekaClient
 
 end